All pages Updated and ready to use. Error Messages has been Added. Now Messaging will be studied.

+++ StudentRegisterSoftware/frmExamEntry.cs
@@ -20,9 +20,9 @@ namespace StudentRegisterSoftware
         public void AvarageCalvulation()
         {
             int x, y, z;
-            x = Convert.ToInt16(txtExamFirst.Text);
-            y = Convert.ToInt16(txtExamSecond.Text);
-            z = Convert.ToInt16(txtExamThird.Text);
+            x = Convert.ToInt16(mskex1.Text);
+            y = Convert.ToInt16(mskex2.Text);
+            z = Convert.ToInt16(mskex3.Text);
 
             avg = (x + y + z) / 3;
         }
@@ -42,9 +42,10 @@ namespace StudentRegisterSoftware
         public void ExamResultClear()
         {
 
-            txtExamFirst.Text = "";
-            txtExamSecond.Text = "";
-            txtExamThird.Text = "";
+            mskex1.Text = "";
+            mskex2.Text = "";
+            mskex3.Text = "";
+            selectedIndex = "";
         }
 
         sqlconnection conn = new sqlconnection();
@@ -59,28 +60,37 @@ namespace StudentRegisterSoftware
         {
             // delete with right click menu
 
-
-            if (restcid == temptcid)
+            if (selectedIndex != "")
             {
-                DialogResult result = new DialogResult();
-                result = MessageBox.Show("Are you sure to DELETE selected data?", "Warning", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
-                if (result == DialogResult.Yes)
+
+
+                if (restcid == temptcid)
                 {
-                    SqlCommand cmd = new SqlCommand("Delete from Tbl_Exams where resid=@p1", conn.conn());
-                    cmd.Parameters.AddWithValue("@p1", selectedIndex);
-                    cmd.ExecuteNonQuery();
-                    conn.conn().Close();
+                    DialogResult result = new DialogResult();
+                    result = MessageBox.Show("Are you sure to DELETE selected data?", "Warning", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+                    if (result == DialogResult.Yes)
+                    {
+                        SqlCommand cmd = new SqlCommand("Delete from Tbl_Exams where resid=@p1", conn.conn());
+                        cmd.Parameters.AddWithValue("@p1", selectedIndex);
+                        cmd.ExecuteNonQuery();
+                        conn.conn().Close();
 
-                    MessageBox.Show("Data has been DELETED?", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        MessageBox.Show("Data has been DELETED?", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
 
-                    ListViaIndexChange("stid=@p1", cmbStudent);
+                        ListViaIndexChange("stid=@p1", cmbStudent);
+                        selectedIndex = "";
+                    }
                 }
+                else
+                {
+                    MessageBox.Show("You can only Delete exam results belong to you", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                }
+
             }
             else
             {
-                MessageBox.Show("You can only Delete exam results belong to you", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                MessageBox.Show("Please Select a Student", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
             }
-
         }
 
         private void copyToBoardToolStripMenuItem_Click(object sender, EventArgs e)
@@ -88,32 +98,23 @@ namespace StudentRegisterSoftware
 
 
             // condition seetings for update
+            if (selectedIndex != "")
+            {
 
+                btnUpdate.Enabled = true;
+                btnSave.Enabled = false;
+                btnSave.BackColor = Color.LightGray;
+                btnUpdate.BackColor = Color.White;
 
-            btnUpdate.Enabled = true;
-            btnSave.Enabled = false;
-            pictureBox4.Visible = true;
-            btnSave.BackColor = Color.LightGray;
-            btnUpdate.BackColor = Color.White;
+                mskex1.Text = examone;
+                mskex2.Text = examtwo;
+                mskex3.Text = examthree;
+            }
+            else
+            {
+                MessageBox.Show("Please Select a Student", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
 
-            txtExamFirst.Text = examone;
-            txtExamSecond.Text = examtwo;
-            txtExamThird.Text = examthree;
-
-
-        }
-
-        private void pictureBox4_Click(object sender, EventArgs e)
-        {
-            //update cancel settings
-            btnSave.Enabled = true;
-            btnUpdate.Enabled = false;
-            btnUpdate.BackColor = Color.LightGray;
-            btnSave.BackColor = Color.White;
-            cmbStudent.Text = "";
-            cmbClass.Text = "";
-            ExamResultClear();
-            pictureBox4.Visible = false;
         }
 
         private void btnExit_Click(object sender, EventArgs e)
@@ -126,68 +127,74 @@ namespace StudentRegisterSoftware
         private void btnUpdate_Click(object sender, EventArgs e)
         {
             // update record
-            try
+
+            if (restcid == temptcid)
             {
+                AvarageCalvulation();
 
-                if (restcid == temptcid)
+                SqlCommand cmd = new SqlCommand("update Tbl_Exams set resexone=@p1, resextwo=@p2, resexthree=@p3, resexavg=@p4, resstatus=@p5 where resid=@p6 and restcid=@p7", conn.conn());
+                cmd.Parameters.AddWithValue("@p1", mskex1.Text);
+                cmd.Parameters.AddWithValue("@p2", mskex2.Text);
+                cmd.Parameters.AddWithValue("@p3", mskex3.Text);
+                cmd.Parameters.AddWithValue("@p4", avg);
+                if (avg < 45)
                 {
-                    AvarageCalvulation();
-
-                    SqlCommand cmd = new SqlCommand("update Tbl_Exams set resexone=@p1, resextwo=@p2, resexthree=@p3, resexavg=@p4, resstatus=@p5 where resid=@p6 and restcid=@p7", conn.conn());
-                    cmd.Parameters.AddWithValue("@p1", txtExamFirst.Text);
-                    cmd.Parameters.AddWithValue("@p2", txtExamSecond.Text);
-                    cmd.Parameters.AddWithValue("@p3", txtExamThird.Text);
-                    cmd.Parameters.AddWithValue("@p4", avg);
-                    if (avg < 45)
-                    {
-                        cmd.Parameters.AddWithValue("@p5", "Failed");
-                    }
-                    else
-                    {
-                        cmd.Parameters.AddWithValue("@p5", "Passed");
-                    }
-
-                    cmd.Parameters.AddWithValue("@p6", selectedIndex);
-                    cmd.Parameters.AddWithValue("@p7", Convert.ToInt16(temptcid));
-
-                    cmd.ExecuteNonQuery();
-                    conn.conn().Close();
-
-                    MessageBox.Show("Exam Results are UPDATED?", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
-
-
-                    ExamResultClear();
-
-                    ListViaIndexChange("stid=@p1", cmbStudent);
-
-                    pictureBox4.Enabled = false;
-                    btnSave.Enabled = true;
-                    btnSave.BackColor = Color.White;
+                    cmd.Parameters.AddWithValue("@p5", "Failed");
                 }
                 else
                 {
-                    MessageBox.Show("You can only change exam results belong to you", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    cmd.Parameters.AddWithValue("@p5", "Passed");
                 }
-            }
-            catch (Exception)
-            {
-                MessageBox.Show("Please provide valid information", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
 
+                cmd.Parameters.AddWithValue("@p6", selectedIndex);
+                cmd.Parameters.AddWithValue("@p7", Convert.ToInt16(temptcid));
+
+                cmd.ExecuteNonQuery();
+                conn.conn().Close();
+
+                MessageBox.Show("Exam Results are UPDATED?", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
+
+
+                ExamResultClear();
+
+                ListViaIndexChange("stid=@p1", cmbStudent);
+
+
+                btnSave.Enabled = true;
+                btnSave.BackColor = Color.White;
             }
+            else
+            {
+                MessageBox.Show("You can only change exam results belong to you", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
+
+
         }
         private void btnSave_Click(object sender, EventArgs e)
         {
             // exam record
-            try
+            int tempvalue = 0;
+
+            SqlCommand cmdselect = new SqlCommand("Select Count (*) from Tbl_Exams where resstid=@p1 and restcid=@p2 ", conn.conn());
+            cmdselect.Parameters.AddWithValue("@p1", cmbStudent.SelectedValue);
+            cmdselect.Parameters.AddWithValue("@p2", temptcid);
+            SqlDataReader dr = cmdselect.ExecuteReader();
+            while (dr.Read())
+            {
+
+                tempvalue = Convert.ToInt16(dr[0].ToString());
+
+            }
+            if(tempvalue == 0)
             {
                 AvarageCalvulation();
 
                 SqlCommand cmd = new SqlCommand("Insert into Tbl_Exams (resstid,restcid,resexone,resextwo,resexthree,resexavg,resstatus) values (@p1,@p2,@p3,@p4,@p5,@p6,@p7)", conn.conn());
                 cmd.Parameters.AddWithValue("@p1", cmbStudent.SelectedValue);
                 cmd.Parameters.AddWithValue("@p2", temptcid);
-                cmd.Parameters.AddWithValue("@p3", txtExamFirst.Text);
-                cmd.Parameters.AddWithValue("@p4", txtExamSecond.Text);
-                cmd.Parameters.AddWithValue("@p5", txtExamThird.Text);
+                cmd.Parameters.AddWithValue("@p3", mskex1.Text);
+                cmd.Parameters.AddWithValue("@p4", mskex2.Text);
+                cmd.Parameters.AddWithValue("@p5", mskex3.Text);
                 cmd.Parameters.AddWithValue("@p6", avg);
                 if (avg < 45)
                 {
@@ -205,12 +212,16 @@ namespace StudentRegisterSoftware
 
                 ExamResultClear();
                 ListViaIndexChange("stid=@p1", cmbStudent);
+                selectedIndex = "";
             }
-            catch (Exception)
+            else
             {
-                MessageBox.Show("Please provide valid information", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                MessageBox.Show("You have already assigned Exam Results for this student.", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
 
             }
+
+
+
         }
 
         private void cmbClass_SelectedIndexChanged(object sender, EventArgs e)
@@ -225,35 +236,22 @@ namespace StudentRegisterSoftware
             cmbStudent.DisplayMember = "Students";
             cmbStudent.DataSource = dt;
             cmbStudent.Text = "";
-            /*
-            // selected class of all students fill
-            SqlCommand cmd2 = new SqlCommand("select resid,(stname+' '+stsurname) as 'STUDENT NAME', (tcname+' '+tcsurname) as 'TEACHER NAME',tcbrans as 'LESSON', resexone as 'EXAM 1', resextwo as 'EXAM 2', resexthree as 'EXAM 3', resexavg as 'EXAM AVERAGE',resstatus as 'STATUS', restcid from Tbl_Exams \r\ninner join Tbl_Students\r\non Tbl_Exams.resstid = Tbl_Students.stid\r\ninner join Tbl_Teachers\r\non Tbl_Exams.restcid = Tbl_Teachers.tcid where stclass=@p1", conn.conn());
-            cmd2.Parameters.AddWithValue("@p1", cmbClass.SelectedValue);
-            SqlDataAdapter da2 = new SqlDataAdapter(cmd2);
-            DataTable dt2 = new DataTable();
-            da2.Fill(dt2);
-            dataGridView1.DataSource = dt2;
-            dataGridView1.Columns[0].Visible = false;
-            */
-
+            ExamResultClear();
+            btnSave.Enabled = true;
+            btnSave.BackColor = Color.White;
             ListViaIndexChange("stclass=@p1", cmbClass);
+            selectedIndex = "";
 
         }
 
         private void cmbStudent_SelectedIndexChanged(object sender, EventArgs e)
         {
-            /*
-             //selected student
-             SqlCommand cmd2 = new SqlCommand("select resid,(stname+' '+stsurname) as 'STUDENT NAME', (tcname+' '+tcsurname) as 'TEACHER NAME',tcbrans as 'LESSON', resexone as 'EXAM 1', resextwo as 'EXAM 2', resexthree as 'EXAM 3', resexavg as 'EXAM AVERAGE' ,resstatus as 'STATUS', restcid from Tbl_Exams \r\ninner join Tbl_Students\r\non Tbl_Exams.resstid = Tbl_Students.stid\r\ninner join Tbl_Teachers\r\non Tbl_Exams.restcid = Tbl_Teachers.tcid where stid=@p1", conn.conn());
-             cmd2.Parameters.AddWithValue("@p1", cmbStudent.SelectedValue);
-             SqlDataAdapter da2 = new SqlDataAdapter(cmd2);
-             DataTable dt2 = new DataTable();
-             da2.Fill(dt2);
-             dataGridView1.DataSource = dt2;
-             dataGridView1.Columns[0].Visible = false;
-            */
-
+            ExamResultClear();
             ListViaIndexChange("stid=@p1", cmbStudent);
+            btnSave.Enabled = true;
+            btnSave.BackColor = Color.White;
+            selectedIndex = "";
+
         }
 
         private void dataGridView1_CellClick(object sender, DataGridViewCellEventArgs e)
@@ -265,8 +263,6 @@ namespace StudentRegisterSoftware
             examtwo = dataGridView1.Rows[choosen].Cells[5].Value.ToString();
             examthree = dataGridView1.Rows[choosen].Cells[6].Value.ToString();
             restcid = dataGridView1.Rows[choosen].Cells[9].Value.ToString();
-
-
         }
 
         private void frmExamEntry_Load(object sender, EventArgs e)
@@ -284,6 +280,7 @@ namespace StudentRegisterSoftware
             cmbClass.DisplayMember = "stclass";
             cmbClass.DataSource = dt;
             cmbClass.Text = "";
+            selectedIndex = "";
